Separate run completion observer
