+++ javatest/javatest-core/src/main/java/io/github/olib963/javatest/TestRunCompletionObserver.java
@@ -1,0 +1,14 @@
+package io.github.olib963.javatest;
+
+import io.github.olib963.javatest.logging.RunLoggingObserver;
+
+@FunctionalInterface
+public interface TestRunCompletionObserver {
+
+    void onRunCompletion(TestResults results);
+
+    static TestRunCompletionObserver logger() {
+        return RunLoggingObserver.INSTANCE;
+    }
+
+}
